Add raw queries to struct2db (#22)
* WIP Add tests for struct2db.DeleteMultiple with raw queries

* Extract some of the struct2db tests to separate files

* Add an struct2db test for getting count with raw query

* Implement _raw and _rawConjuction in struct2db

* Use DeleteMultiple with raw query in the sample_app

+++ pkg/struct2sql/main_test.go
@@ -124,9 +124,9 @@ func TestSQLSelectQueries(t *testing.T) {
 			// However, we need to pass either value or an array so that an array can be extracted into multiple $x's
 			0,
 			0,
-			[]int8{0,0,0,0},
-			[]int8{0,0,0},
-			[]int8{0,0},
+			[]int{0,0,0,0},
+			[]int{0,0,0},
+			[]int{0,0},
 		},
 		"_rawConjuction": RawConjuctionOR,
 	}, map[string]bool{
@@ -170,7 +170,7 @@ func TestSQLDeleteWithFiltersQueries(t *testing.T) {
 			".Price=? OR .EmailSecondary=? OR .Age IN (?)",
 			0,
 			0,
-			[]int8{0,0,0},
+			[]int{0,0,0},
 		},
 	}, map[string]bool{"Price": true})
 	want = "DELETE FROM test_structs WHERE (price=$1) AND (price=$2 OR email_secondary=$3 OR age IN ($4,$5,$6))"
@@ -183,7 +183,7 @@ func TestSQLDeleteWithFiltersQueries(t *testing.T) {
 			".Price=? OR .EmailSecondary=? OR .Age IN (?)",
 			0,
 			0,
-			[]int8{0,0,0},
+			[]int{0,0,0},
 		},
 	}, map[string]bool{"Price": true})
 	want = "DELETE FROM test_structs WHERE (price=$1 OR email_secondary=$2 OR age IN ($3,$4,$5))"
